+++ spec/rails4_dummy_app/app/react/index.js
@@ -1,5 +1,6 @@
-import RWR from 'rwr-integration';
+import RWR from 'react-webpack-rails';
 window.RWR = RWR;
 
 import HelloWorld from './components/hello-world';
 RWR.registerComponent('HelloWorld', HelloWorld);
+